feat(session): add checkpoint infrastructure for mid-flight observability (#1318)
Implement file-based checkpoint system for long-running CSA sessions:
- emit_checkpoint() writes structured TOML checkpoints to session dir
- read_latest_checkpoint() / read_checkpoints() for orchestrator polling
- `csa session checkpoint` CLI subcommand for interactive inspection
- Checkpoints stored as git notes on session commits (no ACP dependency)
